+++ lib/wendigo.ts
@@ -1,11 +1,12 @@
 import process from 'process';
 import puppeteer from 'puppeteer';
+import { BrowserContext, Browser } from './browser/puppeteer_wrapper/puppeteer_types';
 import BrowserFactory from './browser_factory';
 import * as Errors from './errors';
 import { WendigoPluginInterface, BrowserSettings, FinalBrowserSettings, WendigoPluginAssertionInterface, PluginModule } from './types';
 import BrowserInterface from './browser/browser_interface';
 
-const defaultSettings = {
+const defaultSettings: BrowserSettings = {
     log: false,
     headless: true,
     args: [],
@@ -70,17 +71,6 @@ export default class Wendigo {
         BrowserFactory.clearCache();
     }
 
-    // public get Errors(): { [s: string]: typeof Errors.WendigoError | typeof Errors.AssertionError | typeof Errors.TimeoutError } {
-    //     return {
-    //         AssertionError: Errors.AssertionError,
-    //         WendigoError: Errors.WendigoError,
-    //         QueryError: Errors.QueryError,
-    //         FatalError: Errors.FatalError,
-    //         TimeoutError: Errors.TimeoutError,
-    //         InjectScriptError: Errors.InjectScriptError
-    //     };
-    // }
-
     private _validatePlugin(name: string, plugin?: WendigoPluginInterface, assertions?: WendigoPluginAssertionInterface): void {
         this._validatePluginName(name);
         if (plugin && typeof plugin !== 'function') throw new Errors.FatalError("registerPlugin", `Invalid plugin module "${name}".`);
@@ -104,7 +94,7 @@ export default class Wendigo {
         }
     }
 
-    private async _createInstance(settings: FinalBrowserSettings): Promise<puppeteer.BrowserContext | puppeteer.Browser> {
+    private async _createInstance(settings: FinalBrowserSettings): Promise<BrowserContext | Browser> {
         const instance = await puppeteer.launch(settings);
         if (settings.incognito) {
             return instance.createIncognitoBrowserContext();